algorithm/heap.py
class heap(list):
	def __init__(self, data):
		self.__heap = data
		self.__length = len(data)
		self.__heap_size = self.__length
	def left(self, i):
		return 2 * i + 1
	def right(self, i):
		return 2 * i + 2
	def max_heapify(self, i):
		l = self.left(i)
		r = self.right(i)
		if (l <= (self.__heap_size - 1)) and (self.__heap[l] > self.__heap[i]):
			largest = l
		else:
			largest = i
		if (r <= (self.__heap_size - 1)) and (self.__heap[r] > self.__heap[largest]):
			largest = r
		if 	largest != i:
			tmp = self.__heap[i]
			self.__heap[i] = self.__heap[largest]
			self.__heap[largest] = tmp
			self.max_heapify(largest)
	def build_max_heap(self):
		self.__heap_size = self.__length
		for i in range(self.__length / 2 - 1, -1, -1):
			self.max_heapify(i)
	def heapsort(self):
		self.build_max_heap()
		for i in range(self.__length - 1, 0, -1):
			tmp  = self.__heap[0]
			self.__heap[0] = self.__heap[i]
			self.__heap[i] = tmp
			self.__heap_size = self.__heap_size - 1
			self.max_heapify(0)
